+++ src/LLP/corenode.cpp
@@ -152,7 +152,7 @@ namespace LLP
     void CoreNode::ErrorReply(const json::json& jsonError)
     {
         /* Default error status code is 500. */
-        int32_t nStatus = 500;
+        uint16_t nStatus = 500;
 
         /* Get the error code from json. */
         int32_t nError = jsonError["code"].get<int32_t>();
